Extracted MRI2D Segmentation Dataset class
