+++ ivadomed/inference.py
@@ -10,6 +10,8 @@ from loguru import logger
 from torch.utils.data import Dataset
 from torch.utils.data import DataLoader
 from torch import tensor
+
+from ivadomed.loader.mri2d_segmentation_dataset import MRI2DSegmentationDataset
 from ivadomed.transforms import UndoCompose
 from ivadomed import config_manager as imed_config_manager
 from ivadomed import models as imed_models
@@ -362,13 +364,13 @@ def segment_volume(folder_model: str, fname_images: list, gpu_id: int = 0, optio
         logger.info(f"Loaded {len(ds)} {loader_params['slice_axis']} volumes of shape "
                      f"{context['Modified3DUNet']['length_3D']}.")
     else:
-        ds = imed_loader.MRI2DSegmentationDataset(filename_pairs,
-                                                  length=length_2D,
-                                                  stride=stride_2D,
-                                                  slice_axis=slice_axis,
-                                                  cache=True,
-                                                  transform=tranform_lst,
-                                                  slice_filter_fn=imed_loader_utils.SliceFilter(
+        ds = MRI2DSegmentationDataset(filename_pairs,
+                                                                                 length=length_2D,
+                                                                                 stride=stride_2D,
+                                                                                 slice_axis=slice_axis,
+                                                                                 cache=True,
+                                                                                 transform=tranform_lst,
+                                                                                 slice_filter_fn=imed_loader_utils.SliceFilter(
                                                       **loader_params["slice_filter_params"]))
         ds.load_filenames()
         if is_2d_patch:
